Refactored approach: single class per file
