+++ inc/altis-config/collector.php
@@ -1,30 +1,16 @@
 <?php
 
-//namespace Altis\Dev_Tools;
+namespace Altis\Dev_Tools;
 
-use function Altis\get_config;
-//use QM_Collector;
+use QM_Collector;
 
 class QM_Collector_Altis_Config extends QM_Collector {
 
 	public $id = 'altis-config';
 
-	public function __construct() {
-		parent::__construct();
-
-		$this->data['altis-config'] = get_config();
+	function name() {
+		return _x( 'Altis Config', 'Query monitor menu item name', 'altis' );
 	}
-}
 
-/**
- * @param array        $collectors
- * @param QueryMonitor $qm
- *
- * @return array
- */
-function register_qm_collector_altis_config( array $collectors, QueryMonitor $qm ) {
-	$collectors['altis_config'] = new QM_Collector_Altis_Config();
-	return $collectors;
+	public function process() {}
 }
-
-add_filter( 'qm/collectors', 'register_qm_collector_altis_config', 20, 2 );
